requirements.txt
flask==2.3.3
stripe==9.7.0
gspread==5.12.4
google-auth==2.29.0
python-dotenv==1.0.1
